+++ Code/Engine/InputXBox360/InputDeviceXBox.h
@@ -3,31 +3,32 @@
 #include <InputXBox360/Basics.h>
 #include <Core/Input/DeviceTypes/Controller.h>
 
+/// \brief An implementation of ezInputDeviceController that handles XBox 360 controllers.
+///
+/// Works on all platforms that provide the XINPUT API.
 class EZ_INPUTXBOX360_DLL ezInputDeviceXBox360 : public ezInputDeviceController
 {
 public:
   ezInputDeviceXBox360();
 
+  /// \brief Returns an ezInputDeviceXBox360 device.
   static ezInputDeviceXBox360* GetDevice();
+
+  /// \brief Destroys all devices of this type. Automatically called at engine shutdown.
   static void DestroyAllDevices();
 
+  /// \brief Returns "ControllerXBox"
   virtual const char* GetDeviceName() const { return "ControllerXBox"; }
 
-  virtual bool IsControllerConnected(ezUInt8 uiPhysical) const EZ_OVERRIDE { EZ_ASSERT(uiPhysical < 4, "Invalid Controller Index"); return m_bControllerConnected[uiPhysical]; }
-
-  virtual void EnableVibration(ezUInt8 uiController, bool bEnable) EZ_OVERRIDE { m_bEnableVibration[uiController] = bEnable; }
-  virtual bool IsVibrationEnabled(ezUInt8 uiController) const EZ_OVERRIDE { return m_bEnableVibration[uiController]; }
-
-  virtual void SetPhysicalToVirtualControllerMapping(ezUInt8 uiPhysical, ezInt8 iVirtual) EZ_OVERRIDE { m_iMapControllerTo[uiPhysical] = iVirtual; }
-  virtual ezInt8 GetPhysicalToVirtualControllerMapping(ezUInt8 uiPhysical) const EZ_OVERRIDE { return m_iMapControllerTo[uiPhysical]; }
+  virtual bool IsControllerConnected(ezUInt8 uiPhysical) const EZ_OVERRIDE;
 
 private:
+  virtual void ApplyVibration(ezUInt8 uiPhysicalController, Motor::Enum eMotor, float fStrength) EZ_OVERRIDE;
+
   bool m_bControllerConnected[4];
-  ezInt8 m_iMapControllerTo[4];
-  bool m_bEnableVibration[4];
 
   virtual void InitializeDevice() EZ_OVERRIDE { }
-  virtual void UpdateInputSlotValues() EZ_OVERRIDE;
+  virtual void UpdateInputSlotValues(double fTimeDifference) EZ_OVERRIDE;
   virtual void RegisterInputSlots() EZ_OVERRIDE;
 
   void SetValue(ezInt32 iController, const char* szButton, float fValue);