+++ Projects/SampleApp_Game/ShipComponent.cpp
@@ -2,6 +2,7 @@
 #include "ShipComponent.h"
 #include "ProjectileComponent.h"
 #include "CollidableComponent.h"
+#include <InputXBox360/InputDeviceXBox.h>
 
 EZ_IMPLEMENT_COMPONENT_TYPE(ShipComponent, ShipComponentManager);
 
@@ -25,6 +26,9 @@ void ShipComponent::SetVelocity(const ezVec3& vVel)
 
 void ShipComponent::SetIsShooting(bool b)
 {
+  //ezInputDeviceXBox360::GetDevice()->SetVibrationStrength(m_iPlayerIndex, 0, b ? 1.0f : 0.0f);
+  //ezInputDeviceXBox360::GetDevice()->SetVibrationStrength(m_iPlayerIndex, 1, b ? 1.0f : 0.0f);
+
   m_bIsShooting = b;
 }
 
@@ -75,33 +79,41 @@ void ShipComponent::Update()
   if (m_iCurShootCooldown > 0)
     --m_iCurShootCooldown;
   else
-    if (m_bIsShooting && m_iAmmunition >= m_iAmmoPerShot)
+  if (m_bIsShooting && m_iAmmunition >= m_iAmmoPerShot)
+  {
+    m_iCurShootCooldown = m_iShootDelay;
+
+    ezGameObjectDesc desc;
+    desc.m_LocalPosition = m_pOwner->GetLocalPosition();
+    ezGameObjectHandle hProjectile = GetWorld()->CreateObject(desc);
+
+    ezGameObject* pProjectile = GetWorld()->GetObject(hProjectile);
+
+    pProjectile->AddComponent(GetWorld()->GetComponentManager<ProjectileComponentManager>()->CreateComponent());
+
+    ProjectileComponent* pProjectileComponent = pProjectile->GetComponentOfType<ProjectileComponent>();
+    pProjectileComponent->m_iBelongsToPlayer = m_iPlayerIndex;
+    pProjectileComponent->m_vVelocity = vShipDir * ezMath::Max(m_vVelocity.GetLength(), 1.0f) * 1.0f;
+    pProjectileComponent->m_bDoesDamage = true;
+
+    m_iAmmunition -= m_iAmmoPerShot;
+
+    float ShootTrack[20] =
     {
-      m_iCurShootCooldown = m_iShootDelay;
+      1.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f,
+      0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f
+    };
 
-      ezGameObjectDesc desc;
-      desc.m_LocalPosition = m_pOwner->GetLocalPosition();
-      ezGameObjectHandle hProjectile = GetWorld()->CreateObject(desc);
+    ezInputDeviceXBox360::GetDevice()->AddVibrationTrack(m_iPlayerIndex, ezInputDeviceController::Motor::RightMotor, ShootTrack, 20);
+  }
 
-      ezGameObject* pProjectile = GetWorld()->GetObject(hProjectile);
-
-      pProjectile->AddComponent(GetWorld()->GetComponentManager<ProjectileComponentManager>()->CreateComponent());
-
-      ProjectileComponent* pProjectileComponent = pProjectile->GetComponentOfType<ProjectileComponent>();
-      pProjectileComponent->m_iBelongsToPlayer = m_iPlayerIndex;
-      pProjectileComponent->m_vVelocity = vShipDir * ezMath::Max(m_vVelocity.GetLength(), 1.0f) * 1.0f;
-      pProjectileComponent->m_bDoesDamage = true;
-
-      m_iAmmunition -= m_iAmmoPerShot;
-    }
-
-    m_iAmmunition = ezMath::Clamp(m_iAmmunition + 1, 0, m_iMaxAmmunition);
-    m_iHealth = ezMath::Clamp(m_iHealth + 1, 0, 1000);
+  m_iAmmunition = ezMath::Clamp(m_iAmmunition + 1, 0, m_iMaxAmmunition);
+  m_iHealth = ezMath::Clamp(m_iHealth + 1, 0, 1000);
 
 
-    ezVec3 vCurPos = m_pOwner->GetLocalPosition();
-    vCurPos = vCurPos.CompMax(ezVec3(-20.0f));
-    vCurPos = vCurPos.CompMin(ezVec3(20.0f));
+  ezVec3 vCurPos = m_pOwner->GetLocalPosition();
+  vCurPos = vCurPos.CompMax(ezVec3(-20.0f));
+  vCurPos = vCurPos.CompMin(ezVec3(20.0f));
 
-    m_pOwner->SetLocalPosition(vCurPos);
+  m_pOwner->SetLocalPosition(vCurPos);
 }